WIP: Expose a FIPS indicator
FIPS 140-3 requires us to indicate whether an operation was using
approved services or not. The FIPS 140-3 implementation guidelines
provide two basic approaches to doing this: implicit indicators, and
explicit indicators.

Implicit indicators are basically the concept of "if the operation
passes, it was approved". We were originally aiming for implicit
indicators in our copy of OpenSSL. However, this proved to be a problem,
because we wanted to certify a signature service, and FIPS 140-3
requires that a signature service computes the digest to be signed
within the boundaries of the FIPS module. Since we were planning to
certify fips.so only, this means that EVP_PKEY_sign/EVP_PKEY_verify
would have to be blocked. Unfortunately, EVP_SignFinal uses
EVP_PKEY_sign internally, but outside of fips.so and thus outside of the
FIPS module boundary. This means that using implicit indicators in
combination with certifying only fips.so would require us to block both
EVP_PKEY_sign and EVP_SignFinal, which are the two APIs currently used
by most users of OpenSSL for signatures.

EVP_DigestSign would be acceptable, but has only been added in 3.0 and
is thus not yet widely used.

As a consequence, we've decided to introduce explicit indicators so that
EVP_PKEY_sign and EVP_SignFinal can continue to work for now, but
FIPS-aware applications can query the explicit indicator to check
whether the operation was approved.

This patch is currently WIP and only adds an indicator using
EVP_MD_fips_indicator() for digests, but the same concept should expand
to all other uses of algorithms implemented in the FIPS module.

The simplest approach to tagging each algorithm implementation with its
associated FIPS approval status would be extending OSSL_ALGORITHM and
OSSL_DISPATCH. This would, however, modify ABI between OpenSSL core and
the providers, while this approach of defining a separate struct and
a separate query function does not. Modules that do not implement the
new query function would just end up being marked as unapproved.

The fips indicator status flag is obtained during algorithm fetching if
available and eventually stored in the EVP_MD struct (or, when extended
to other use cases, in the corresponding user-visible handle).

Signed-off-by: Clemens Lang <cllang@redhat.com>